Implemented collision tests within a class so they can be re-used by a gtest
